add linux arm64 targets
This adds a new ubuntu arm64 runner to the matrix and compiles
the app and the cli there.
For the linux arm64 app, we add the NO_STRIP=true env variable,
according to a known limitiation in tauri using linuxdeploy to create appimages
https://github.com/linuxdeploy/linuxdeploy/issues/272
The *-musl targets are still cross-compiled with zig, because I could
not test that it would work without zig.

close #50